requirements.txt
inference==0.50.3
numpy==2.2.6
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
Pillow==11.2.1
python-dotenv==1.1.0
